pyproject.toml
[project]
name = "hatch-conda-build"
description = 'conda-build using pyproject.toml'
dynamic = ["version"]
requires-python = ">=3.8"
readme = "README.md"
dependencies = [
   "grayskull",
   "hatchling",
   "ruamel.yaml",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "hatch_conda_build/__version__.py"

[project.entry-points.hatch]
conda = "hatch_conda_build.hooks"

[tool.hatch.envs.test]
dependencies = [
  "build",
  "ruff",
  "black",
  "pytest",
]

[tool.hatch.envs.test.scripts]
check = [
  "black --check --diff .",
  "ruff check .",
]
format = [
  "black .",
]
test = "pytest"